feat: set description for VIP 352
