+++ simulations/vip-352/simulations.ts
@@ -1,0 +1,7 @@
+import { forking, testVip } from "src/vip-framework";
+
+import vip352 from "../../vips/vip-352/bscmainnet";
+
+forking(41186665, async () => {
+  testVip("VIP-352", await vip352(), {});
+});
